moved group endpoints to aaa (#339)
* moved getgroups, grouplist to aaa

* added boolean success flags
