+++ src/ai/susi/server/api/aaa/GetGroupDetails.java
@@ -1,4 +1,4 @@
-package ai.susi.server.api.cms;
+package ai.susi.server.api.aaa;
 
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
@@ -13,7 +13,7 @@ import java.util.Set;
  * Created by chetankaushik on 15/06/17.
  * This servlet returns details of the group whose name is passed in the URL Get parameters.
  * This accepts one GET parameter, which is the group name you want to search for.
- * This can be tested on http://127.0.0.1:4000/cms/getGroupDetails.json?group={groupName}
+ * This can be tested on http://127.0.0.1:4000/aaa/getGroupDetails.json?group={groupName}
  *
  */
 public class GetGroupDetails extends AbstractAPIHandler implements APIHandler {
@@ -33,7 +33,7 @@ public class GetGroupDetails extends AbstractAPIHandler implements APIHandler {
 
     @Override
     public String getAPIPath() {
-        return "/cms/getGroupDetails.json";
+        return "/aaa/getGroupDetails.json";
     }
 
     @Override
@@ -41,6 +41,7 @@ public class GetGroupDetails extends AbstractAPIHandler implements APIHandler {
         Boolean foundUser;
         JSONObject success = new JSONObject();
         success.put("accepted", false);
+        success.put("message", "Error: Unable to process request");
         JSONObject allUsers;
         allUsers = DAO.group.toJSON();
         String model_name = call.get("group", null);
